Add config file support
