+++ samples/sample10/resource/application1.properties
@@ -1,0 +1,1 @@
+property1="test1"
