+++ samples/sample10/resource/application2.properties
@@ -1,0 +1,1 @@
+property2="test2"
