+++ samples/sample10/resource/ballerina.conf
@@ -1,0 +1,7 @@
+# username->userid mapping
+[john]
+userid="john@ballerina.com"
+groups="apim,esb"
+[jane]
+userid="jane3@ballerina.com"
+groups="esb"
